http: Fix bug in connection_lost.
The connection_lost handler of the HTTP protocol feeds self._error
into the queue, but that is only set by the super-class
implementation, which is called afterwards. Enqueue the exception
object that is passed to the super-class implementation instead.

This problem causes the caller to receive a HttpResponse object with
_message = None.