http: Fix re-exporting of symbols from http_client again
Starting with python 2.7.10, httplib contains _METHODS_EXPECTING_BODY,
which is a set. The name matches the heuristic used to find constants
but breaks with the `in` check because the set cannot be hashed.

This commit refactors the re-export of the http status constants to
hopefully be more robust. It iterates over the `responses` dictionary
and tries to derive the names of the constants from the values.

This new approach currently covers all entries of the responses
dictionary that have a corresponding constant.